phy: rockchip-inno-usb2: Turn on the USB PHY in DCP mode
Booting a rv1108-elgin board without the USB cable connected and
connecting the USB cable afterwards causes the g_serial driver
to not be detected in the host PC.

The reason for the non-detection is that the USB PHY is turned off
in DCP mode.

Fix it by turning it on so that the USB host can detect g_serial reliably.

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>