ARM: dts: imx6dl-omni700: Use PLL5 as LDB clock source
The et070wblg3 display requires a pixel clock of 33.28MHz.

Currently the LDB clock source is the default MMDC_CH1 and provides
a pixel clock of 28.28MHz, which is outside the range of the display
specification.

Use PLL5 as the LDB clock source, which can provide an accurate 33.28MHz
as required.

Signed-off-by: Fabio Estevam <festevam@gmail.com>