media: i2c: adv7180: Allow the control of the reset pin
On a design where the ADV7180 pin is pulled down, it is not possible
to make it functional unless the GPIO connected to this pin goes
high.

Add support for the reset GPIO by introducing an optional property
called 'reset-gpios'.

Note: the reset operation is still performed via software as recommended
by the Analog Devices, but the reset GPIO still needs to go high to make
the chip operational.

Signed-off-by: Fabio Estevam <festevam@gmail.com>