media: i2c: adv7180: Don't enable CSI2 output before streaming started
Currently the MIPI CSI data output is turned on when s_power() is
called. This can cause trouble with the host, when the upstream
controller is turned on while the data stream is already running.

Let's move the switching of the output to s_stream() to handle this
properly.

TODO: Fix the runtime suspend

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>
[fabio: protected the csi_write() calls with ADV7180_FLAG_MIPI_CSI2 dependency]
Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>