media: i2c: adv7180: Validate input channel upon probe
The ADV7280-M has the reset value of 0x0e in register
ADV7180_REG_INPUT_CONTROL which selects the differential CVBS
input that is not even supported on this chip.

Therefore we validate the reset value of the register against
the valid_input_mask and if a mismatch occurs we select the
first input channel by setting the register to 0x00.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>
Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>